% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6. power_divergence_functions.R
\name{A_diff2_g}
\alias{A_diff2_g}
\title{The A function in the split of the GP density differentiated two times to gamma}
\usage{
A_diff2_g(v, A)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape third parameter the observations}
}
\value{
the derivative in the given point v
}
\description{
The A function in the split of the GP density differentiated two times to gamma
}
