% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6. power_divergence_functions.R
\name{A_func}
\alias{A_func}
\title{The A function in the split of the GP density}
\usage{
A_func(v)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape third parameter the observations}
}
\value{
the function evaluated at v
}
\description{
The A function in the split of the GP density
}
