% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. CV_grid_search.R
\name{add_folds_togehter}
\alias{add_folds_togehter}
\title{Add the cross validation of different folds together}
\usage{
add_folds_togehter(fold_results)
}
\arguments{
\item{fold_results}{list of results of each fold}
}
\value{
a pointwise mean over the different folds
}
\description{
Add the cross validation of different folds together
}
