% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6. power_divergence_functions.R
\name{B_diff_s}
\alias{B_diff_s}
\title{The B function in the split of the GP density differentiated to sigma}
\usage{
B_diff_s(v, A, B)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape third parameter the observations}
}
\value{
the derivative in the given point v
}
\description{
The B function in the split of the GP density differentiated to sigma
}
