% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.partial_dependence.R
\name{calc_PD}
\alias{calc_PD}
\title{Calculate partial dependence for gbex object}
\usage{
calc_PD(object, var_name)
}
\arguments{
\item{object}{A gbex object}
}
\value{
A list with for each variable the partial dependence of beta and gamma with a
numeric value corresponding to the variable.
}
\description{
Calculate partial dependence for gbex object
}
