% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9. variable_importance.R
\name{calc_VI}
\alias{calc_VI}
\title{Calculate variable importance for a gbex object}
\usage{
calc_VI(object, type = c("relative", "permutation"))
}
\arguments{
\item{object}{A gbex object}

\item{type}{a character vector either relative for "relative" importance or "permutation" for permutation importance}
}
\value{
A named vector with scaled importance scores for each variable
}
\description{
Calculate variable importance for a gbex object
}
\details{
The relative importance score is calculated sepearately for sigma and gamma parameter.
the permutation importance measure is calculated for deviance.
}
