% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. CV_tuning_parameters.R
\name{CV_depth}
\alias{CV_depth}
\title{Cross validation for the deth of the trees}
\usage{
CV_depth(y, X, num_folds, depth_list, Bmax, stratified, ...)
}
\arguments{
\item{y}{Numeric vector of response}

\item{X}{Data frame of covariates}

\item{num_folds}{the number of folds used to determine the optimal number of trees}

\item{depth_list}{A list with parameter values for depth}

\item{Bmax}{maximum number of trees used for finding the optimal}

\item{stratified}{indicate whether stratified sampling should be used}

\item{...}{Additional arguments supplied to gbex function}
}
\value{
A CV_gbex object
}
\description{
Cross validation for the deth of the trees
}
