% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_gbex_B.R
\name{CV_gbex_B}
\alias{CV_gbex_B}
\title{Cross validation for the number of trees B}
\usage{
CV_gbex_B(y, X, num_folds, Bmax, stratified, ncores, ...)
}
\arguments{
\item{y}{numeric vector of response}

\item{X}{data.frame of covariates}

\item{num_folds}{numeric number of folds to use}

\item{Bmax}{numeric maximum number of trees to fit for each fold}

\item{stratified}{boolean indicating to use stratified sampling for the folds}

\item{ncores}{number of cores to used for parallelization of the cross validation, if not specified cores are chosen by detectCores function from the parallel package (Does not work on windows)}

\item{...}{Other arguments to be passed to the gbex function (details)}
}
\value{
A CV_gbex object see description of CV_gbex
}
\description{
Cross validation for the number of trees B
}
