% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_gbex_par.R
\name{CV_gbex_par}
\alias{CV_gbex_par}
\title{Cross validation for a parameter in gbex function}
\usage{
CV_gbex_par(y, X, num_folds, par_name, par_grid, Bmax, stratified, ncores, ...)
}
\arguments{
\item{y}{numeric vector of response}

\item{X}{data.frame of covariates}

\item{num_folds}{numeric number of folds to use}

\item{par_name}{character name of the parameter for which to perform a cross validation (see details of CV_gbex)}

\item{par_grid}{list or vector a grid for the parameter in par_name (see details of CV_gbex)}

\item{Bmax}{numeric maximum number of trees to fit for each fold}

\item{stratified}{boolean indicating to use stratified sampling for the folds}

\item{ncores}{number of cores to used for parallelization of the cross validation, if not specified cores are chosen by detectCores function from the parallel package (Does not work on windows)}

\item{...}{Other arguments to be passed to the gbex function (details)}
}
\value{
A CV_gbex object see CV_gbex
}
\description{
Cross validation for a parameter in gbex function
}
