% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. CV_grid_search.R
\name{CV_grid_search}
\alias{CV_grid_search}
\title{Cross Validation Grid Search}
\usage{
CV_grid_search(y, X, par_name = NULL, grid = NULL, num_folds = 4,
  par_fixed = list())
}
\arguments{
\item{y}{Numeric vector of response}

\item{X}{Data frame of covariates}

\item{par_name}{Character name of the parameter to adjust}

\item{grid}{Either a numeric vector or list of possible parameter values}

\item{num_folds}{Integer indicating the number of folds}

\item{par_fixed}{Named list of values of parameters that are kept fixed}
}
\value{
The function returns a named list with the following elements:
\item{par}{Optimal parameter value for all parameter values in grid}
\item{par_name}{The paramter name}
\item{grid}{A vector or list with tried parameter values}
\item{dev}{Numeric vector with deviance for each parameter value in grid}
\item{B}{Integers with optimal B value for each value in grid}
}
\description{
For a given grid for a parameter find the optimal value such that the cross validation deviance is minimized
}
