% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. CV_tuning_parameters.R
\name{CV_lambda_ratio}
\alias{CV_lambda_ratio}
\title{Cross validation for the ratio of lambda_sigma and lambda_gamma}
\usage{
CV_lambda_ratio(y, X, num_folds, lambda_vec, Bmax, stratified, ...)
}
\arguments{
\item{y}{Numeric vector of response}

\item{X}{Data frame of covariates}

\item{num_folds}{the number of folds used to determine the optimal number of trees}

\item{lambda_vec}{A list with parameter values for lambda_ratio}

\item{Bmax}{maximum number of trees used for finding the optimal}

\item{stratified}{indicate whether stratified sampling should be used}

\item{...}{Additional arguments supplied to gbex function}
}
\value{
A CV_gbex object
}
\description{
Cross validation for the ratio of lambda_sigma and lambda_gamma
}
