% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. CV_grid_search.R
\name{CV_model_run}
\alias{CV_model_run}
\title{Run the gradient boosting algorithm for a specific set of parameters}
\usage{
CV_model_run(Xtrain, ytrain, Xtest, ytest, B, lambda, depth, min_leaf_size,
  sf)
}
\arguments{
\item{Xtrain}{data.frame of covariates for training}

\item{ytrain}{vector of response for training}

\item{Xtest}{data.frame of covariates for testing}

\item{ytest}{vector of response for testing}

\item{B}{maximum number of trees}

\item{lambda}{vector with learning rates for sigma and gamma}

\item{depth}{vector with depth of the trees for sigma and gamma}

\item{min_leaf_size}{vector with the minimum leaf size of trees for sigma and gamma}

\item{sf}{the sample fraction for each tree}
}
\value{
the deviance for each step of the boosting algorithm
}
\description{
Run the gradient boosting algorithm for a specific set of parameters
}
