% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. CV_tuning_parameters.R
\name{CV_normal}
\alias{CV_normal}
\title{Cross validation to inspect performance of the  model}
\usage{
CV_normal(y, X, num_folds, Bmax, stratified, ncores, ...)
}
\arguments{
\item{y}{Numeric vector of response}

\item{X}{Data frame of covariates}

\item{num_folds}{the number of folds used to determine the optimal number of trees}

\item{Bmax}{maximum number of trees used for finding the optimal}

\item{stratified}{indicate whether stratified sampling should be used}

\item{ncores}{number of cores to use for parallelization, if not specified cores are chosen by detectCores function from the parallel package}

\item{...}{Additional arguments supplied to gbex function}
}
\value{
A CV_gbex object
}
\description{
Cross validation to inspect performance of the  model
}
