% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. CV_grid_search.R
\name{CV_single_fold}
\alias{CV_single_fold}
\title{Estimate a single fold for cross validation}
\usage{
CV_single_fold(fold, folds, data, args)
}
\arguments{
\item{fold}{Integer indicating which fold should be computed}

\item{folds}{Integer vector of length n with the division in folds}

\item{data}{List where response in named y and covariates are name X}

\item{args}{Named list with arguments for the gbex model}
}
\value{
numeric vector with deviance of each boosting iteration
}
\description{
Estimate a single fold for cross validation
}
