% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1. gradient_boosting_functions.R
\name{CV_tuning_parameters}
\alias{CV_tuning_parameters}
\title{Tuning parameters}
\usage{
CV_tuning_parameters(y, X, num_folds = 8, lambda_grid, lambda, Bmax,
  par_fixed, par_grid = NULL)
}
\arguments{
\item{y}{Numeric vector of observations}

\item{X}{Data frame with the right column names}

\item{num_folds}{Integer for the number of folds in the cross validation}

\item{lambda_grid}{Numeric with learning rate for sigma and gamma used for doing the grid search}

\item{lambda}{Numeric with learning rate for sigma and gamma used for final model}

\item{Bmax}{Numeric indicating the maximum number of trees}

\item{par_fixed}{Named list with for each parameter an initial value}

\item{par_grid}{Named list with for each parameter a grid to optimize over}
}
\value{
A named list with the optimal parameters
}
\description{
Obtain optimal tuning parameters for gbex by performing a grid search.
}
\details{
The function uses two lambda parameters.
The parameters lambda_grid should be larger than lambda and is used to perform the gridsearch as less trees need to be fitted.
}
