% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1. gradient_boosting_functions.R
\name{dev_per_step}
\alias{dev_per_step}
\title{Deviance per step}
\usage{
dev_per_step(object, y = NULL, X = NULL)
}
\arguments{
\item{object}{A fitted gbex object}

\item{y}{A vector of observations}

\item{X}{A dataframe with the right column names}
}
\value{
A vector with the deviance at each step
}
\description{
Compute the deviance per boosting step
}
