% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9. gpd_functions.R
\name{dgpd}
\alias{dgpd}
\title{density function GPD}
\usage{
dgpd(x, sigma, gamma)
}
\arguments{
\item{x}{A probability level}

\item{sigma}{a vector of sigma values}

\item{gamma}{a vecotr of gamma values =}
}
\value{
density of the GPD
}
\description{
density function GPD
}
