% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. CV_tuning_parameters.R
\name{divide_in_folds}
\alias{divide_in_folds}
\title{Divide the data up in folds for cross validation}
\usage{
divide_in_folds(y, num_folds, stratified = F)
}
\arguments{
\item{y}{Vector of observations}

\item{num_folds}{Integer number of folds used for cross validation}

\item{stratified}{Boolean indicating wheter stratified sampling should be used}
}
\value{
A vector with integers corresponding to the fold
}
\description{
Divide the data up in folds for cross validation
}
\details{
The stratified sampling creates stratified clusters of size num_folds from the ordered observations y.
For each fold one observations is sampled.
}
