% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fill_leafs_g}
\alias{fill_leafs_g}
\title{Fill the leafnodes of the gamma tree using linesearch}
\usage{
fill_leafs_g(tree, DF_tree)
}
\arguments{
\item{tree}{a gradient tree}

\item{DF_tree}{the boosting dataframe used for obtaining the tree}
}
\value{
A data.frame object with for each leaf the gradient update
}
\description{
Fill the leafnodes of the gamma tree using linesearch
}
