% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_test_functions.R
\name{find_optimal_B}
\alias{find_optimal_B}
\title{Function to find optimal number of trees with other parmaeters given}
\usage{
find_optimal_B(y, X, num_folds = 4, Bmax, lambda, depth, min_leaf_size,
  sf)
}
\arguments{
\item{y}{Numeric vector of response}

\item{X}{Data frame of covariates}

\item{num_folds}{the number of folds used to determine the optimal number of trees}

\item{Bmax}{maximum number of trees used for finding the optimal}

\item{lambda}{the learning rate}

\item{depth}{the depth of the trees}

\item{min_leaf_size}{the minimum leaf node size}

\item{sf}{subample fraction used for fitting the trees}
}
\value{
The optimal number of trees used
}
\description{
Function to find optimal number of trees with other parmaeters given
}
