% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9. threshold_function.R
\name{forest_threshold}
\alias{forest_threshold}
\title{Estimate a threshold model for gbex extrapolation}
\usage{
forest_threshold(y, X, p = 0.8, ...)
}
\arguments{
\item{y}{Response variables (vector of length n)}

\item{X}{Covariate matrix (matrix of dimension (n x d))}

\item{p}{probability level to estimate the threshold}

\item{...}{additional arguments for the quantile forest}
}
\value{
a list with the estimated quantile forest and the estimated threshold
}
\description{
Estimate a threshold model for gbex extrapolation
}
