% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. CV_grid_search.R
\name{get_argument_list}
\alias{get_argument_list}
\title{Get a list of function arguments for gbex for each of the grid_values}
\usage{
get_argument_list(data, par, grid_values, par_fixed)
}
\arguments{
\item{data}{a list with training and testing data}

\item{par}{a character with the name of the paramter over which to do a grid search}

\item{grid_values}{a numeric vector with values over which to do a grid_search}

\item{par_fixed}{a named list with fixed values for the other parameters}
}
\value{
a list with all arguments for the gbex function
}
\description{
Get a list of function arguments for gbex for each of the grid_values
}
