% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. CV_grid_search.R
\name{get_CV_data}
\alias{get_CV_data}
\title{Get a list of training and testing data for a single fold}
\usage{
get_CV_data(fold, folds, X, y)
}
\arguments{
\item{fold}{a numeric indicating which of the folds in the cross validation}

\item{folds}{a vector with for each data point the fold to which they belong}

\item{X}{data.frame of covariates}

\item{y}{vector of response}
}
\value{
a list with training and testing data
}
\description{
Get a list of training and testing data for a single fold
}
