% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. CV_grid_search.R
\name{get_CV_df}
\alias{get_CV_df}
\title{Calculate the CV dataframe with for each parameter value the deviance and the best chosen number of trees}
\usage{
get_CV_df(performance, grid_values, par)
}
\arguments{
\item{performance}{a matrix as returned by performance_by_fold}

\item{grid_values}{the values of the parameter to try}

\item{par}{character vecotr with the name of a single parameter}
}
\value{
A dataframe with for each parameter value the deviance and the best chosen number of trees
}
\description{
Calculate the CV dataframe with for each parameter value the deviance and the best chosen number of trees
}
