% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8. data_simulation_functions.R
\name{get_data}
\alias{get_data}
\title{Simulate a dataset}
\usage{
get_data(n, d)
}
\arguments{
\item{n}{number of observations}

\item{d}{dimension of the covariate space}
}
\value{
the second derivative of the negative log likelihood to gamma for a single observation
}
\description{
Simulate a dataset
}
