% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_DF_boost}
\alias{get_DF_boost}
\title{Create or update the DF_boost data.frame}
\usage{
get_DF_boost(DF, theta)
}
\arguments{
\item{DF}{A data frame with the response and covariates or when updating the current DF_boost data frame}

\item{theta}{the new parameter updates for sigma and gamma}
}
\value{
A data.frame object with the data the parameter estimates the negative likelihood and the first and second derivatives of the negative likelihood towards sigma and gamma
}
\description{
Create or update the DF_boost data.frame
}
