% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_leaf_values}
\alias{get_leaf_values}
\title{For a gradient tree fill the leafnodes with the gradient update gamma}
\usage{
get_leaf_values(tree)
}
\arguments{
\item{tree}{a gradient tree}
}
\value{
A data.frame object with for each leaf the gradient update
}
\description{
For a gradient tree fill the leafnodes with the gradient update gamma
}
