% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. CV_grid_search.R
\name{get_optimal_parameters}
\alias{get_optimal_parameters}
\title{Calculate the optimal parameter}
\usage{
get_optimal_parameters(CV_df)
}
\arguments{
\item{CV_df}{a dataframe obtained by get_CV_df}
}
\value{
a numeric vector with the optimal parameters
}
\description{
Calculate the optimal parameter
}
