% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. CV_tuning_parameters.R
\name{GoF}
\alias{GoF}
\title{Goodness of fit statistics transformed random variables}
\usage{
GoF(U, type)
}
\arguments{
\item{U}{A vector of transformed random variables}

\item{type}{the teststatistic to compute choose from ("CvM","AD") for cramer van mises and anderson darling}
}
\value{
the value of the teststatistic
}
\description{
Goodness of fit statistics transformed random variables
}
