% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5. power_divergence_functions.R
\name{GP_density}
\alias{GP_density}
\title{The density of the generalized pareto distribution}
\usage{
GP_density(v, A, B)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape third parameter the observations}
}
\value{
the density in point v
}
\description{
The density of the generalized pareto distribution
}
