% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6. power_divergence_functions.R
\name{GP_density_diff_g}
\alias{GP_density_diff_g}
\title{The density of the generalized pareto distribution differentiated to gamma}
\usage{
GP_density_diff_g(v, A, B)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape third parameter the observations}
}
\value{
the density in point v
}
\description{
The density of the generalized pareto distribution differentiated to gamma
}
