% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6. likelihood_functions.R
\name{GP_dev}
\alias{GP_dev}
\title{The negative log likelihood for the GPD}
\usage{
GP_dev(v)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape and the third entry the observation}
}
\value{
the negative log likelihood for a single observation
}
\description{
The negative log likelihood for the GPD
}
