% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. likelihood_functions.R
\name{GP_dev_diff2_s}
\alias{GP_dev_diff2_s}
\title{The negative log likelihood for the GPD two times differantiated to sigma}
\usage{
GP_dev_diff2_s(v)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape and the third entry the observation}
}
\value{
the second derivative of the negative log likelihood to sigma for a single observation
}
\description{
The negative log likelihood for the GPD two times differantiated to sigma
}
