% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. likelihood_functions.R
\name{GP_dev_diff_g}
\alias{GP_dev_diff_g}
\title{The negative log likelihood for the GPD differantiated to gamma}
\usage{
GP_dev_diff_g(v)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape and the third entry the observation}
}
\value{
the derivative of the negative log likelihood to gamma for a single observation
}
\description{
The negative log likelihood for the GPD differantiated to gamma
}
