% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9. gpd_functions.R
\name{gpd_crps_loss}
\alias{gpd_crps_loss}
\title{crps loss for gpd estimates}
\usage{
gpd_crps_loss(sigma, gamma, y)
}
\arguments{
\item{sigma}{a vector of sigma values}

\item{gamma}{a vecotr of gamma values}

\item{y}{a vector of observations}
}
\value{
crps loss for the gpd
}
\description{
crps loss for gpd estimates
}
