% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9. gpd_functions.R
\name{gpd_quant_loss}
\alias{gpd_quant_loss}
\title{quantile loss for gpd estimates}
\usage{
gpd_quant_loss(sigma, gamma, y, tau)
}
\arguments{
\item{sigma}{a vector of sigma values}

\item{gamma}{a vecotr of gamma values}

\item{y}{a vector of observations}

\item{tau}{a probability level for the quantile loss}
}
\value{
quantile loss for the tau probability level
}
\description{
quantile loss for gpd estimates
}
