% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. gradient_trees.R
\name{gradient_tree}
\alias{gradient_tree}
\title{Fit gradient Tree}
\usage{
gradient_tree(X, rr, rr2, depth, min_leaf_size)
}
\arguments{
\item{X}{the covariates a data.frame}

\item{rr}{the derivative of likelihood or powerdivergence}

\item{rr2}{the second derivative of likelihood or powerdivergence}

\item{depth}{A value indicating the maximum depth}

\item{min_leaf_size}{A value indicating the minimum leafsize}
}
\value{
A list with tree object and a named vector with for each leafnode the update value
}
\description{
Estimate a gradient tree then calculate for the leafnodes the optimal gradient based on a single newton raphson step
}
