% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. gradient_trees.R
\name{gradient_tree_beta}
\alias{gradient_tree_beta}
\title{Gradient Tree beta}
\usage{
gradient_tree_beta(tree_df, depth, min_leaf_size)
}
\arguments{
\item{tree_df}{A dataframe to estimate the tree of the same format as the boosting_df}

\item{depth}{A value indicating the maximum depth}

\item{min_leaf_size}{A value indicating the minimum leafsize}
}
\value{
A list with tree object and a named vector with for each leafnode the update value
}
\description{
Estimate a tree for the beta variable then calculate for the leafnodes how to update.
Beta is a reparametrization of sigma in the sense of beta = log(sigma)
}
