% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6. power_divergence_functions.R
\name{Int}
\alias{Int}
\title{The integral in the first part of empirical power divergence}
\usage{
Int(v)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape and the third parameter the alpha}
}
\value{
the integral for parameter vector v
}
\description{
The integral in the first part of empirical power divergence
}
