% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5. power_divergence_functions.R
\name{Int_diff2_s}
\alias{Int_diff2_s}
\title{The integral in the first part of empirical power divergence two times differentiated to sigma}
\usage{
Int_diff2_s(v)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape and the third parameter the alpha}
}
\value{
the derivative in point v
}
\description{
The integral in the first part of empirical power divergence two times differentiated to sigma
}
