% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. CV_tuning_parameters.R
\name{loss_per_step}
\alias{loss_per_step}
\title{Loss per step}
\usage{
loss_per_step(object, y = NULL, X = NULL, loss_func = "likelihood", tau = 0.5)
}
\arguments{
\item{object}{A fitted gbex object}

\item{y}{A vector of observations}

\item{X}{A dataframe with the right column names}

\item{loss_func}{the loss function used}

\item{tau}{the probability level for quantile loss (only used when loss_func = "quant_loss")}
}
\value{
A vector with the loss at each step
}
\description{
Compute the loss function per boosting step
}
