% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_compute.R
\name{parallel_compute}
\alias{parallel_compute}
\title{Compute list in parallel}
\usage{
parallel_compute(l, func, vars, envir, ncores)
}
\arguments{
\item{l}{elements to parallize}

\item{func}{function to apply on elements of l}

\item{vars}{list of variable names to be distributed over the nodes}

\item{envir}{environment in which the varibales live}

\item{ncores}{number of cores to parallize}
}
\value{
A CV_gbex object see description of CV_gbex
}
\description{
Compute list in parallel
}
