% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependence.R
\name{partial_dependence}
\alias{partial_dependence}
\title{Partial dependence function}
\usage{
partial_dependence(object, var_name, grid_points = 50)
}
\arguments{
\item{object}{A gbex object}

\item{var_name}{character variable names for which to calculate the partial dependence function (maximum number of variables is 2)}

\item{grid_points}{numetic number of equally spaced grid points for a single dimension}
}
\value{
A list with for each variable the partial dependence of sigma and gamma with a
numeric value corresponding to the variable.
}
\description{
Partial dependence function
}
\details{
For a two dimensional partial dependence function the number of grid points used is grid_points^2.
}
