% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependence.R
\name{partial_dependence_quantile}
\alias{partial_dependence_quantile}
\title{Partial dependence function for quantiles}
\usage{
partial_dependence_quantile(
  object,
  tau,
  var_name,
  grid_points = 50,
  grid = NULL
)
}
\arguments{
\item{object}{A gbex object}

\item{var_name}{character variable names for which to calculate the partial dependence function (maximum number of variables is 2)}

\item{grid_points}{numeric number of equally spaced grid points for a single dimension}

\item{grid}{named dataframe with grid on which to compute partial dependence (default NULL)}
}
\value{
A list with for each variable the partial dependence of sigma and gamma with a
numeric value corresponding to the variable.
}
\description{
Partial dependence function for quantiles
}
\details{
For a two dimensional partial dependence function the number of grid points used is grid_points^2.
}
