% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6. power_divergence_functions.R
\name{PD_dev}
\alias{PD_dev}
\title{The density power divergence}
\usage{
PD_dev(v, A, B, alpha = 0.5)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape and the third entry the observation}

\item{alpha}{the power for the power divergence}
}
\value{
Power divergence evaluated in the point v
}
\description{
The density power divergence
}
