% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_divergence_functions.R
\name{PD_dev_diff2_g}
\alias{PD_dev_diff2_g}
\title{The density power divergence differentiated two times to gamma}
\usage{
PD_dev_diff2_g(v, A, B, alpha = 0.5)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape and the third entry the observation}

\item{alpha}{the power for the power divergence}
}
\value{
the derivative in the point v
}
\description{
The density power divergence differentiated two times to gamma
}
