% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5. power_divergence_functions.R
\name{PD_dev_diff2_s}
\alias{PD_dev_diff2_s}
\title{The density power divergence differentiated two times to sigma}
\usage{
PD_dev_diff2_s(v, A, B, alpha = 0.5)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape and the third entry the observation}

\item{alpha}{the power for the power divergence}
}
\value{
the derivative in point v
}
\description{
The density power divergence differentiated two times to sigma
}
