% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.partial_dependence.R
\name{PD_tree}
\alias{PD_tree}
\title{Calculate partial dependence for gradient tree}
\usage{
PD_tree(tree, var_name, values)
}
\arguments{
\item{tree}{A gradient_tree object}

\item{var_name}{variable name for which to calulate the partial dependence}

\item{values}{Numeric vector with the values for var_name to calculate partial dependence}
}
\value{
A vector with partial dependence from the gradient tree for vector values
}
\description{
Calculate partial dependence for gradient tree
}
\details{
The implementation makes use of the structure of the trees.
By keeping the observations in each node the tree can be followed down one time for the computation.
}
