% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. CV_grid_search.R
\name{performance_per_fold}
\alias{performance_per_fold}
\title{Calculate deviance per fold in the cross validation}
\usage{
performance_per_fold(arg_lists)
}
\arguments{
\item{arg_lists}{a list with argument lists to call the gbex function}
}
\value{
a matrix where each row matches the the elements of arg_lists and the columns the number of trees
}
\description{
Calculate deviance per fold in the cross validation
}
