% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9. gpd_functions.R
\name{pgpd}
\alias{pgpd}
\title{Distribution function GPD}
\usage{
pgpd(x, sigma, gamma)
}
\arguments{
\item{x}{A probability level}

\item{sigma}{a vector of sigma values}

\item{gamma}{a vecotr of gamma values =}
}
\value{
distribution of the GPD
}
\description{
Distribution function GPD
}
