% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. CV_plotting_function.R
\name{plot.CV_gbex}
\alias{plot.CV_gbex}
\title{Plotting function for CV_par object}
\usage{
\method{plot}{CV_gbex}(object, what = "general")
}
\arguments{
\item{object}{par_CV object}

\item{what}{string taking "general" average over all folds, "per_fold" for a figure for each fold or "hist_data" for histograms of the different folds.}
}
\value{
plots the cross validation results for the CV_gbex object
}
\description{
Plotting function for CV_par object
}
