% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbex.R
\name{plot.gbex}
\alias{plot.gbex}
\title{Plot function for gbex}
\usage{
\method{plot}{gbex}(object)
}
\arguments{
\item{object}{A fitted gbex object}
}
\value{
Plots the deviance for each iteration
}
\description{
Plot function for gbex
}
