% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_partial_dependence.R
\name{plot.PD_gbex}
\alias{plot.PD_gbex}
\title{Plot function for PD_gbex}
\usage{
\method{plot}{PD_gbex}(object)
}
\arguments{
\item{object}{A PD_gbex object}
}
\value{
Two figures side by side with on the left the partial dependence plot for sigma and on the right the partial dependence plot for gamma.
}
\description{
Plot function for PD_gbex
}
