% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1. gradient_boosting_functions.R
\name{predict.gbex}
\alias{predict.gbex}
\title{Predict function for gbex object}
\usage{
\method{predict}{gbex}(object, newdata = NULL)
}
\arguments{
\item{object}{A fitted gbex object}

\item{newdata}{A data frame with covariates for which to predict the sigma and gamma parameter}
}
\value{
A data.frame object with the estimated sigma and gamma parameters
}
\description{
Predict function for gbex object
}
