% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. gradient_trees.R
\name{predict.gradient_tree}
\alias{predict.gradient_tree}
\title{Get the gradient update for each leafnode}
\usage{
\method{predict}{gradient_tree}(object, newdata = NULL)
}
\arguments{
\item{object}{A gradient tree object}

\item{newdata}{A New data frame with columns used for the gradient tree; If NULL the updates for the model data is computed}
}
\value{
A vector with gradient steps
}
\description{
Get the gradient update for each leafnode
}
