% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. CV_plotting_function.R
\name{print.CV_gbex}
\alias{print.CV_gbex}
\title{printing function for CV_gbex object}
\usage{
\method{print}{CV_gbex}(object)
}
\arguments{
\item{object}{par_CV object}
}
\value{
prints a basic description of the CV_gbex object
}
\description{
printing function for CV_gbex object
}
