% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbex.R
\name{print.gbex}
\alias{print.gbex}
\title{Print function for gbex}
\usage{
\method{print}{gbex}(object)
}
\arguments{
\item{object}{A fitted gbex object}
}
\value{
Prints the gbex object
}
\description{
Print function for gbex
}
