% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9. gpd_functions.R
\name{qgpd}
\alias{qgpd}
\title{Quantile function GPD}
\usage{
qgpd(p, sigma, gamma)
}
\arguments{
\item{p}{A probability level}

\item{sigma}{a vector of sigma values}

\item{gamma}{a vecotr of gamma values =}
}
\value{
quantile of the GPD
}
\description{
Quantile function GPD
}
