% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7. data_simulation_functions.R
\name{sim_gpd_data}
\alias{sim_gpd_data}
\title{Simulate data from gpd distribution}
\usage{
sim_gpd_data(n, s, g)
}
\arguments{
\item{n}{number of observations}

\item{s}{sigma parameter for each observation}

\item{g}{gamma parameter for each observation}
}
\value{
simulated data from gpd with specified parameter
}
\description{
Simulate data from gpd distribution
}
