% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2. boosting_preperation.R
\name{transform_parameters}
\alias{transform_parameters}
\title{Transform parameters to constrain them to be positive or not}
\usage{
transform_parameters(theta, gamma_positive, inverse_transform = F)
}
\arguments{
\item{theta}{data frame with parameters}

\item{gamma_positive}{boolean indicating whether gamma should be positive}

\item{inverse_transform}{boolean indicating wheter the tranformation should be inverted}
}
\value{
tranformed parameters
}
\description{
Transform parameters to constrain them to be positive or not
}
