% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1. gradient_boosting_functions.R
\name{trim_trees}
\alias{trim_trees}
\title{Trim trees from gbex object}
\usage{
trim_trees(object, B_new)
}
\arguments{
\item{object}{A fitted gbex object}

\item{B_new}{A new total number of trees}
}
\value{
A gbex object with B_new trees
}
\description{
Trim trees from gbex object
}
