% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{update_parameter}
\alias{update_parameter}
\title{Update the current parameters by one gradient step}
\usage{
update_parameter(TREE_s, TREE_g, DF_boost, lambda)
}
\arguments{
\item{TREE_s}{a list of gradient tree for sigma together with a df with gradient step per leafnode}

\item{DF_boost}{the current boosting data frame}

\item{lambda}{a vector of length two with learning rate for sigma and gamma}

\item{TREE_s}{a list of gradient tree for sigma together with a df with gradient step per leafnode}
}
\value{
A data.frame object with for each leaf the gradient update
}
\description{
Update the current parameters by one gradient step
}
