% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8. variable_importance.R
\name{variable_importance}
\alias{variable_importance}
\title{Calculate variable importance for a gbex object}
\usage{
variable_importance(object, type = "relative")
}
\arguments{
\item{object}{A gbex object}
}
\value{
A plot with an ordered histogram of the scaled variable importance measures
}
\description{
Calculate variable importance for a gbex object
}
\details{
In the case of relative importance two figures are supplied with the importance for sigma and for gamma.
For permutation importance only one figure is supplied as importance for deviance.
}
