% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9. variable_importance.R
\name{VI_permutation}
\alias{VI_permutation}
\title{Calculate permutation variable importance}
\usage{
VI_permutation(var_name, object)
}
\arguments{
\item{var_name}{The variable name}

\item{object}{A gbex object}
}
\value{
A named vector with importance scores relative to deviance
}
\description{
Calculate permutation variable importance
}
