% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VI_relative.R
\name{VI_relative}
\alias{VI_relative}
\title{Relative variable importance for a gradient tree}
\usage{
VI_relative(tree, var_names)
}
\arguments{
\item{tree}{a gradient tree object}

\item{var_names}{The set of variable names used for fitting}
}
\value{
A named vector with importance scores of the variables for the tree
}
\description{
Relative variable importance for a gradient tree
}
