% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9. variable_importance.R
\name{VI_relative_tree}
\alias{VI_relative_tree}
\title{Calculate relative variable importance for a gradient tree}
\usage{
VI_relative_tree(tree, var_names)
}
\arguments{
\item{tree}{a gradient tree object}

\item{var_names}{The set of variable names used for fitting}
}
\value{
A named vector with importance scores of the variables for the tree
}
\description{
Calculate relative variable importance for a gradient tree
}
