% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8. variable_importance.R
\name{calc_VI}
\alias{calc_VI}
\title{Calculate variable importance for a gbex object}
\usage{
calc_VI(object, type = c("relative", "permutation"), scaled = T)
}
\arguments{
\item{object}{A gbex object}

\item{type}{a character vector either relative for "relative" importance or "permutation" for permutation importance}

\item{scaled}{boolean indicating wheteher the variable importance should be scaled between 0-100 or to use the raw importance}
}
\value{
A named vector with scaled importance scores for each variable
}
\description{
Calculate variable importance for a gbex object
}
\details{
The relative importance score is calculated sepearately for sigma and gamma parameter.
the permutation importance measure is calculated for deviance.
}
