% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2. boosting_preperation.R
\name{compute_divergence}
\alias{compute_divergence}
\title{Divergence}
\usage{
compute_divergence(y, theta, gamma_positive)
}
\arguments{
\item{y}{Vector of observations}

\item{theta}{Named data frame of parameters for the generelized pareto distribution}

\item{gamma_positive}{boolean indicating whether gamma should be positive}
}
\value{
A data frame GPD parameters calculated power divergence and derivatives of power divergence
}
\description{
Computes the power divergence for the generalized pareto distribution together with the first and second derivatives.
}
\details{
the sigma parameter is reparematerized as exp(beta) hence the output are the derivatives in terms of beta and gamma
}
