% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. CV_grid_search.R
\name{CV_grid_search}
\alias{CV_grid_search}
\title{Optimize the model by doing a gridsearch over a parameter using cross validation}
\usage{
CV_grid_search(X, y, par = NULL, grid_values = NULL, num_folds = 4,
  par_fixed)
}
\arguments{
\item{X}{data.frame of covariates}

\item{y}{vector of response}

\item{par}{character name of the parameter to adjust, if it is NULL cross validation is performed using parameters in par_fixed}

\item{grid_values}{numeric vector of values to perform gridsearch ove, if it is NULL cross validation is performed using parameters in par_fixed}

\item{num_folds}{integer number of folds used for cross validation}

\item{par_fixed}{named list of values of parameters that should remain fixed}
}
\value{
the optimal parameter value and a dataframe with cross validation performance for each of the parameter values together with the chosen B
}
\description{
Optimize the model by doing a gridsearch over a parameter using cross validation
}
