% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_boosting_functions.R
\name{dev_per_step}
\alias{dev_per_step}
\title{Compute the deviance at each iteration for a given dataset}
\usage{
dev_per_step(object, X = NULL, y = NULL)
}
\arguments{
\item{object}{A fitted gbex object}

\item{X}{A dataframe with the right column names}

\item{y}{A vector of observations}
}
\value{
A vector with the deviance at iteration
}
\description{
Compute the deviance at each iteration for a given dataset
}
