% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. CV_grid_search.R
\name{divide_in_folds}
\alias{divide_in_folds}
\title{Divide the data up in folds for cross validation}
\usage{
divide_in_folds(y, num_folds)
}
\arguments{
\item{y}{vector of response}

\item{num_folds}{integer number of folds used for cross validation}
}
\value{
a vector with integers corresponding to the fold
}
\description{
Divide the data up in folds for cross validation
}
