% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. CV_grid_search.R
\name{divide_in_folds}
\alias{divide_in_folds}
\title{Divide the data up in folds for cross validation}
\usage{
divide_in_folds(n, num_folds)
}
\arguments{
\item{n}{Integer number of observations}

\item{num_folds}{Integer number of folds used for cross validation}
}
\value{
A vector with integers corresponding to the fold
}
\description{
Divide the data up in folds for cross validation
}
