% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_in_folds.R
\name{divide_in_folds}
\alias{divide_in_folds}
\title{Divide observations randomly over folds}
\usage{
divide_in_folds(y, num_folds, stratified = F)
}
\arguments{
\item{num_folds}{Integer number of folds used for cross validation}

\item{stratified}{Boolean indicating wheter sampling should be done in a stratified way}

\item{n}{numeric number of observations}
}
\value{
A vector with integers corresponding to the fold
}
\description{
Function assigns randomly assigns each observation to a fold for cross validation
}
\details{
The stratified sampling creates stratified clusters of size num_folds from the ordered observations y.
For each fold one observations is sampled.
}
