% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fill_leafs_s}
\alias{fill_leafs_s}
\title{Fill the leafnodes of the sigma tree using linesearch}
\usage{
fill_leafs_s(tree, DF_boost)
}
\arguments{
\item{tree}{a gradient tree}

\item{DF_boost}{the boosting data.frame}
}
\value{
A data.frame object with for each leaf the gradient update
}
\description{
Fill the leafnodes of the sigma tree using linesearch
}
