% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2. boosting_preperation.R
\name{first_guess}
\alias{first_guess}
\title{Make the first guess of GPD parameters}
\usage{
first_guess(y, gamma_positive)
}
\arguments{
\item{y}{A vector of observations}

\item{gamma_positive}{boolean indicating whether gamma should be positive}
}
\value{
Returning a data frame with in each of the length(y) rows the estimated parameters.
}
\description{
Make the first guess of GPD parameters
}
\details{
Estimation is done based on unconditional maximum likelihood.
Where the density is reparameterized by beta = log(sigma).
}
