% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2. helper_functions.R
\name{first_guess}
\alias{first_guess}
\title{Estimate the unconditional tail as a first guess}
\usage{
first_guess(y)
}
\arguments{
\item{y}{A set of observations}

\item{newdata}{A data frame with covariates for which to predict the sigma and gamma parameter}
}
\value{
A data.frame object with the estimated sigma and gamma parameters
}
\description{
Estimate the unconditional tail as a first guess
}
