% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_guess.R
\name{first_guess}
\alias{first_guess}
\title{Make the first guess of GPD parameters}
\usage{
first_guess(y, gamma_positive)
}
\arguments{
\item{y}{numeric vector of positive observations}

\item{gamma_positive}{boolean indicating whether gamma should be positive for boosting procedure}
}
\value{
Returning a data.frame with length(y) identical rows with the estimated parameters for sigma and gamma.
}
\description{
Estimate parameters of a GPD distribution on a set of observations
}
\details{
The maximum likelihood estimator is used for the to obtain the parameters.
When gamma < 0 and gamma_positive= True gamma is set to 0.01 and a warning is issued.
}
