% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_boosting_functions.R
\name{gbex}
\alias{gbex}
\title{Estimate scale and shape for the GPD using gradient boosting}
\usage{
gbex(y, X, B = 180, lambda = c(0.025, 0.0025), depth = c(2, 2),
  min_leaf_size = c(30, 30), sf = 0.5, alpha = 0)
}
\arguments{
\item{y}{Response variables (vector of length n)}

\item{X}{Covariate matrix (matrix of dimension (n x d))}

\item{B}{Number of gradient boosting steps}

\item{lambda}{learning rate for the scale and shape parameter}

\item{depth}{Maximum depth of the trees}

\item{sf}{sample fraction used for fitting the trees}

\item{alpha}{the power for power divergence (default alpha = 0 meaning maximum likelihood is used)}
}
\value{
A list with the estimates of scale and shape the deviance for each optimaization step a list of trees and a data frame with all parameters of the final optimization step.
}
\description{
Estimate scale and shape for the GPD using gradient boosting
}
