% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbex.R
\name{gbex}
\alias{gbex}
\title{GPD boosting}
\usage{
gbex(
  y,
  X,
  B = 100,
  lambda = NULL,
  lambda_ratio = 7,
  lambda_scale = 0.05,
  depth = c(2, 2),
  min_leaf_size = rep(max(10, length(y)/100), 2),
  sf = 0.75,
  gamma_positive = T,
  initial_values = NULL,
  silent = F,
  pred_sigma = NULL,
  pred_gamma = NULL
)
}
\arguments{
\item{y}{Response variables (vector of length n)}

\item{X}{Covariate matrix (matrix of dimension (n x d))}

\item{B}{numeric Number of gradient boosting steps}

\item{lambda}{numeric, learning rate for the sigma and gamma parameter}

\item{lambda_ratio}{numeric, the ratio of the lambda_sigma/lambda_gamma (Use this together with lambda_scale, see details)}

\item{lambda_scale}{numeric learning rate of lambda_sigma (Use this together with lambda_ratio, see details)}

\item{depth}{numeric, the maximum depth for the sigma and gamma trees}

\item{min_leaf_size}{numeric, of length two indicatin the minimum of observations in a leaf node for sigma and gamma trees.}

\item{sf}{numeric, sample fraction between 0 and 1}

\item{initial_values}{numeric, the initial values of the sigma and gamma parameter. If NULL the parameters are estimated by fitting the GPD to Y.}

\item{silent}{boolean if true progress is given during the fitting procedure.}

\item{pred_sigma}{charcter vector with names of the predictors to use for sigma (default NULL all covariates will be used)}

\item{pred_gamma}{charcter vector with names of the predictors to use for gamma (default NULL all covariates will be used)}
}
\value{
gbex returns an object of class "gbex" which contains the following components:
\item{theta}{Data frame with the estimated gamma and sigma parameter for each observation}
\item{dev}{Numeric with deviance of model}
\item{theta_init}{numeric with inital parameters for sigma and gamma}
\item{gamma_positive}{boolean indicating if the model is fit restricting the gamma parameter to be positive}
\item{trees_sigma}{List with gradient_tree objects for sigma}
\item{trees_gamma}{List with gradient_tree objects for gamma}
\item{lambda}{Numeric with the learining rate used for sigma and gamma}
\item{B}{Numeric with number of trees}
\item{depth}{Numeric with maximum tree depth for sigma and gamma}
\item{sf}{Numeric subsample fraction used in fitting}
\item{min_leaf_size}{numeric minimum number of observations per leaf used in fitting}
\item{data}{The data used in fitting}
}
\description{
gbex is used to fit the parameters of a generalized pareto distribution as a function of covariates to data.
In a boosting procedure two sequences of trees are fitted, one for the sigma parameter (scale) and one for the gamma parameter (shape).
}
\details{
The learning rates of the two two sequence of trees are interconnected, therefore it is more natural to only define the learning rate for the scale, and
indirectly specifying the other lambda by specifying lambda_ratio = lambda_sigma/lambda_gamma.
The tuning parameters can be chosen by means of cross validation via CV_gbex.
}
