% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2. boosting_preperation.R
\name{get_boosting_df}
\alias{get_boosting_df}
\title{Boosting DF}
\usage{
get_boosting_df(data, theta, alpha = 0)
}
\arguments{
\item{data}{A data frame with the response and covariates}

\item{theta}{The parameter values for each observation}

\item{alpha}{The power for the divergence if 0 then maximum likelihood is used (default 0)}
}
\value{
A data frame with the observations, parameter values and derivatives of the power divergence
}
\description{
Create an data frame with the observations, current value of the parameters and the derivitives of the power divergence.
Used to perform one step the the boosting procedure.
}
