% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boosting_df.R
\name{get_boosting_df}
\alias{get_boosting_df}
\title{Boosting DF}
\usage{
get_boosting_df(data, theta, gamma_positive)
}
\arguments{
\item{data}{A data frame with the response and covariates}

\item{theta}{The parameter values for each observation}

\item{gamma_positive}{boolean indicating whether gamma should be positive}
}
\value{
A data frame with the data parameter values and derivatives of the GPD likelihood
}
\description{
Given the data current parameter estimates update compute the likelihood and derivatives of the parameters
for the next tree to be fitted in the boosting procedure of gbex.
}
