% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_dev.R
\name{GP_dev}
\alias{GP_dev}
\alias{GP_dev_diff_s}
\alias{GP_dev_diff_g}
\alias{GP_dev_diff2_s}
\alias{GP_dev_diff2_g}
\title{Likelihood functions for the generalized pareto density}
\usage{
GP_dev(v)

GP_dev_diff_s(v)

GP_dev_diff_g(v)

GP_dev_diff2_s(v)

GP_dev_diff2_g(v)
}
\arguments{
\item{v}{vector with first entry the scale second entry shape and the third entry the observation}
}
\value{
the negative log likelihood for a single observation
}
\description{
Likelihood functions for the generalized pareto density
}
