% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_tree.R
\name{gradient_tree}
\alias{gradient_tree}
\title{Fit gradient tree}
\usage{
gradient_tree(X, rr, rr2, depth, min_leaf_size)
}
\arguments{
\item{X}{the covariates a data.frame}

\item{rr}{the derivative of likelihood or powerdivergence}

\item{rr2}{the second derivative of likelihood or powerdivergence}

\item{depth}{A value indicating the maximum depth}

\item{min_leaf_size}{A value indicating the minimum leafsize}
}
\value{
A gradient_tree object containing
\item{tree}{The fitted tree object of class rpart}
\item{update_table}{data.frame with for each leafnode a newton raphson optimization step}
}
\description{
Fit a gradient tree and fill the leaf nodes with updates based on a newton raphson optimization step
}
\details{
The procedure relies on the rpart function of the rpart package.
Only here the leafnode predictions are not average of rr of observations falling into the leaf, but instead a newton raphson optimization
step.
}
