% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.partial_dependence.R
\name{partial_dependence}
\alias{partial_dependence}
\title{Make partial dependence plot for given variable}
\usage{
partial_dependence(object, variable)
}
\arguments{
\item{object}{A gbex object}

\item{variable}{variable name or index for which to make the dependence plot}
}
\value{
Two figures side by side with on the left the partial dependence plot for beta and on the right the partial dependence plot for gamma.
}
\description{
Make partial dependence plot for given variable
}
