% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD_tree.R
\name{PD_tree}
\alias{PD_tree}
\alias{PD_tree2}
\title{Calculate partial dependence for gradient tree}
\usage{
PD_tree(tree, var_name, values)

PD_tree2(tree, var_names, values)
}
\arguments{
\item{tree}{A gradient_tree object}

\item{var_name}{variable name for which to calulate the partial dependence}

\item{values}{Numeric vector with the values for var_name to calculate partial dependence}

\item{var_names}{variable names for which to calculate the partial dependence}
}
\value{
A vector with partial dependence from the gradient tree for vector values
}
\description{
Calculate partial dependence for gradient tree
}
\details{
The implementation makes use of the structure of the trees.
By keeping the observations in each node the tree can be followed down one time for the computation.
The function PD_tree is used for calculating one dimensional partial dependence functions and PD_tree2 is used for 2 dimensional partial dependence functions.
}
