% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CV_gbex.R
\name{plot.CV_gbex}
\alias{plot.CV_gbex}
\title{Plot function for CV_gbex}
\usage{
\method{plot}{CV_gbex}(object)
}
\arguments{
\item{object}{par_CV object}
}
\value{
linegraph of the the cross validation deviance for the different parameter values in the par_grid
}
\description{
Plot function for CV_gbex
}
