% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gbex.R
\name{predict.gbex}
\alias{predict.gbex}
\title{Predict function for gbex}
\usage{
\method{predict}{gbex}(object, newdata = NULL, probs = NULL, what = "par", Blim = NULL)
}
\arguments{
\item{object}{A fitted gbex object}

\item{newdata}{A data.frame with covariates for which to predict if NULL it is done for insample data}

\item{probs}{numeric with probabilities for which to estimate quantiles  (needs what == "quant")}

\item{what}{character with "par" for parameters, "quant" for quantiles}

\item{Blim}{Numeric number of trees to use for the prediction}
}
\value{
A data.frame object with the estimated sigma and gamma parameters or a data.frame with the estimated quantiles
}
\description{
Fpredicted values for a fitted gbe model.
}
