% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gradient_tree.R
\name{predict.gradient_tree}
\alias{predict.gradient_tree}
\title{Predict the gradient tree}
\usage{
\method{predict}{gradient_tree}(object, newdata = NULL)
}
\arguments{
\item{object}{A gradient tree object}

\item{newdata}{A New data frame with columns used for the gradient tree; If NULL the updates for the model data is computed}
}
\value{
A vector with gradient steps
}
\description{
predicting the gradient updates for the parameter
}
