% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_CV_gbex.R
\name{print.CV_gbex}
\alias{print.CV_gbex}
\title{Print function for CV_gbex}
\usage{
\method{print}{CV_gbex}(object)
}
\arguments{
\item{object}{CV_gbex object}
}
\value{
prints a basic description of the CV_gbex object
}
\description{
Print function for CV_gbex
}
