% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_parameters.R
\name{transform_parameters}
\alias{transform_parameters}
\title{Transform parameters for fitting procedure}
\usage{
transform_parameters(theta, gamma_positive, inverse_transform = F)
}
\arguments{
\item{theta}{data.frame with parameters named s and g and for transformed paramaters st and gt}

\item{gamma_positive}{boolean indicating whether gamma should be positive}

\item{inverse_transform}{boolean indicating wheter the tranformation should be inverted}
}
\value{
tranformed parameters
}
\description{
Transform parameters for fitting procedure
}
\details{
In the parameter the
The transformation is done based on parameter change where for example sigma = exp(sigma_transformed).
The trasformation is from sigma_transformed to sigma and the inverse transformation is from the other way around.
}
