% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_parameters.R
\name{update_parameters}
\alias{update_parameters}
\title{Update parameters}
\usage{
update_parameters(tree_sigma, tree_gamma, boosting_df, lambda)
}
\arguments{
\item{tree_sigma}{A gradient tree object fitted for sigma parameter}

\item{tree_gamma}{A gradient tree object fitted for gamma parameter}

\item{boosting_df}{the boosting_df}

\item{lambda}{a vector with the learning rate of sigma and gamma}
}
\value{
A data frame with the udpated parameters sigma and gamma
}
\description{
Function is used in the boosting procedure of gbex to update the parameters after a set of gradient trees is fitted.
}
