% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{variable_importance}
\alias{variable_importance}
\title{Variable importance for a gbex object}
\usage{
variable_importance(object, type = c("relative", "permutation"), scaled = T)
}
\arguments{
\item{object}{A gbex object}

\item{type}{either "relative" for relative importance or "permutation" for permutation importancee}

\item{scaled}{boolean if True the variable importance is scaled such that the most important variable is scaled to 100.}
}
\value{
A VI_gbex object with
\item{type}{the type of variable importance}
\item{VI_df}{data.frame with the variable importance measures for each variable}
\item{scaled}{If the variable importance is scaled after computing}
}
\description{
Variable importance for a gbex object
}
\details{
The relative importance score is calculated sepearately for sigma and gamma parameter and the permutation importance measure is calculated for deviance.
}
