% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VI_permutation.R
\name{VI_permutation}
\alias{VI_permutation}
\title{Permutation variable importance}
\usage{
VI_permutation(var_name, object)
}
\arguments{
\item{var_name}{Variable name for which to calculate variable importance}

\item{object}{A gbex object}
}
\value{
A named vector with importance scores in decrease of deviance
}
\description{
Permutation variable importance
}
